The price can be changed by the user (for demonstration purposes). Check whether the quantity of stock purchased is allowed, set the quantity data to double.

+++ src/main/java/entity/CommonSimulatedHoldingFactory.java
@@ -5,7 +5,7 @@ package entity;
  */
 public class CommonSimulatedHoldingFactory implements SimulatedHoldingFactory {
     @Override
-    public SimulatedHolding create(String symbol, double price, int amount) {
+    public SimulatedHolding create(String symbol, double price, double amount) {
         return new CommonSimulatedHolding(symbol, price, amount);
     }
 }
